# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sou Rifeiro API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /viagens:
    get:
      operationId: listViagens
      tags: [viagens]
      summary: List all trips
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [planejada, ativa, concluida, cancelada]
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Viagem"
    post:
      operationId: createViagem
      tags: [viagens]
      summary: Create a new trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateViagemBody"
      responses:
        "201":
          description: Created trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Viagem"
  /viagens/{id}:
    get:
      operationId: getViagem
      tags: [viagens]
      summary: Get a single trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Viagem"
        "404":
          description: Not found
    patch:
      operationId: updateViagem
      tags: [viagens]
      summary: Update a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateViagemBody"
      responses:
        "200":
          description: Updated trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Viagem"
        "404":
          description: Not found
    delete:
      operationId: deleteViagem
      tags: [viagens]
      summary: Delete a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /viagens/{id}/occupancy:
    get:
      operationId: getViagemOccupancy
      tags: [viagens]
      summary: Get trip occupancy stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Occupancy stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ViagemOccupancy"

  # ── Clientes (Customers) ───────────────────────────
...
  /produtos:
    get:
      operationId: listProdutos
      tags: [produtos]
      summary: List all products
      parameters:
        - name: categoriaId
          in: query
          schema:
            type: integer
        - name: ativo
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Produto"
    post:
      operationId: createProduto
      tags: [produtos]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProdutoBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produto"
  /produtos/{id}:
    get:
      operationId: getProduto
      tags: [produtos]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produto"
        "404":
          description: Not found
    patch:
      operationId: updateProduto
      tags: [produtos]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProdutoBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produto"
        "404":
          description: Not found
    delete:
      operationId: deleteProduto
      tags: [produtos]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Categorias ─────────────────────────────────────
...
components:
  schemas:
...
    Viagem:
      type: object
      required: [id, destino, dataPartida, dataRetorno, assentosTotal, assentosVendidos, precoPorAssento, status, createdAt]
      properties:
        id:
          type: integer
        destino:
          type: string
        descricao:
          type: ["string", "null"]
        dataPartida:
          type: string
          format: date-time
        dataRetorno:
          type: string
          format: date-time
        assentosTotal:
          type: integer
        assentosVendidos:
          type: integer
        precoPorAssento:
          type: number
        paradasRota:
          type: ["string", "null"]
        status:
          type: string
          enum: [planejada, ativa, concluida, cancelada]
        imagemUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    CreateViagemBody:
      type: object
      required: [destino, dataPartida, dataRetorno, assentosTotal, precoPorAssento]
      properties:
        destino:
          type: string
        descricao:
          type: string
        dataPartida:
          type: string
          format: date-time
        dataRetorno:
          type: string
          format: date-time
        assentosTotal:
          type: integer
        precoPorAssento:
          type: number
        paradasRota:
          type: string
        status:
          type: string
          enum: [planejada, ativa, concluida, cancelada]
        imagemUrl:
          type: string
    UpdateViagemBody:
      type: object
      properties:
        destino:
          type: string
        descricao:
          type: string
        dataPartida:
          type: string
          format: date-time
        dataRetorno:
          type: string
          format: date-time
        assentosTotal:
          type: integer
        precoPorAssento:
          type: number
        paradasRota:
          type: string
        status:
          type: string
          enum: [planejada, ativa, concluida, cancelada]
        imagemUrl:
          type: string
    ViagemOccupancy:
      type: object
      required: [viagemId, destino, assentosTotal, assentosVendidos, assentosDisponiveis, percentualOcupacao, receitaTotal, receitaPendente]
      properties:
        viagemId:
          type: integer
        destino:
          type: string
        assentosTotal:
          type: integer
        assentosVendidos:
          type: integer
        assentosDisponiveis:
          type: integer
        percentualOcupacao:
          type: number
        receitaTotal:
          type: number
        receitaPendente:
          type: number

    # ── Cliente Schemas ────────────────────────────────
...
    Produto:
      type: object
      required: [id, nome, preco, ativo, createdAt]
      properties:
        id:
          type: integer
        nome:
          type: string
        descricao:
          type: ["string", "null"]
        preco:
          type: number
        imagemUrl:
          type: ["string", "null"]
        categoriaId:
          type: ["integer", "null"]
        categoriaNome:
          type: ["string", "null"]
        ativo:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateProdutoBody:
      type: object
      required: [nome, preco]
      properties:
        nome:
          type: string
        descricao:
          type: string
        preco:
          type: number
        imagemUrl:
          type: string
        categoriaId:
          type: integer
        ativo:
          type: boolean
    UpdateProdutoBody:
      type: object
      properties:
        nome:
          type: string
        descricao:
          type: string
        preco:
          type: number
        imagemUrl:
          type: string
        categoriaId:
          type: integer
        ativo:
          type: boolean

    # ── Categoria Schemas ──────────────────────────────
...