# lib/api-zod/src/generated/types/createViagemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sou Rifeiro API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateViagemBodyStatus } from "./createViagemBodyStatus";

export interface CreateViagemBody {
  destino: string;
  descricao?: string;
  dataPartida: Date;
  dataRetorno: Date;
  assentosTotal: number;
  precoPorAssento: number;
  paradasRota?: string;
  status?: CreateViagemBodyStatus;
  imagemUrl?: string;
}